+++ ruleminer/evaluator.py
@@ -56,12 +56,7 @@ class CodeEvaluator:
         self._mean_logs = []
         self._std_logs = []
         self._quantile_logs = []
-        self._eq_logs = []
-        self._ne_logs = []
-        self._ge_logs = []
-        self._le_logs = []
-        self._gt_logs = []
-        self._lt_logs = []
+        self._eval_logs = []
 
     def set_globals(self):
         """
@@ -244,28 +239,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side == right_side
             else:
@@ -281,7 +256,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     "==",
-                    self._eq_logs,
                 )
                 return (max_left >= min_right) & (min_left <= max_right)
 
@@ -294,28 +268,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side == right_side
             else:
@@ -334,28 +288,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side <= right_side
             else:
@@ -372,28 +306,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side <= right_side
             else:
@@ -409,7 +323,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     "<=",
-                    self._le_logs,
                 )
                 return min_left <= max_right
 
@@ -422,28 +335,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side < right_side
             else:
@@ -460,28 +353,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side < right_side
             else:
@@ -497,7 +370,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     "<",
-                    self._le_logs,
                 )
                 return min_left < max_right
 
@@ -510,28 +382,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side >= right_side
             else:
@@ -548,28 +400,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side >= right_side
             else:
@@ -585,7 +417,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     ">=",
-                    self._ge_logs,
                 )
                 return max_left >= min_right
 
@@ -598,28 +429,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side > right_side
             else:
@@ -636,28 +447,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side > right_side
             else:
@@ -673,7 +464,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     ">",
-                    self._ge_logs,
                 )
                 return max_left > min_right
 
@@ -686,28 +476,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side != right_side
             else:
@@ -723,7 +493,6 @@ class CodeEvaluator:
                     min_right,
                     max_right,
                     "!=",
-                    self._ne_logs,
                 )
                 return ~((max_left >= min_right) & (min_left <= max_right))
 
@@ -736,28 +505,8 @@ class CodeEvaluator:
             right_side_neg,
         ):
             """ """
-            if (
-                any(
-                    [
-                        p(left_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
-            ) or (
-                any(
-                    [
-                        p(right_side)
-                        for p in [
-                            pd.api.types.is_string_dtype,
-                            pd.api.types.is_bool_dtype,
-                            pd.api.types.is_datetime64_ns_dtype,
-                        ]
-                    ]
-                )
+            if (self.datatype_not_apply_xbrl_tolerance(left_side)) or (
+                self.datatype_not_apply_xbrl_tolerance(right_side)
             ):
                 return left_side != right_side
             else:
@@ -908,6 +657,18 @@ class CodeEvaluator:
         self.globals["_div"] = _div
         self.globals["_corr"] = _corr
 
+    def datatype_not_apply_xbrl_tolerance(self, value):
+        return any(
+            [
+                p(value)
+                for p in [
+                    pd.api.types.is_string_dtype,
+                    pd.api.types.is_bool_dtype,
+                    pd.api.types.is_datetime64_ns_dtype,
+                ]
+            ]
+        )
+
     def _log(
         self,
         left_side,
@@ -917,172 +678,184 @@ class CodeEvaluator:
         min_right,
         max_right,
         operator,
-        logger,
     ):
         """ """
         # {left-right=diff} operator [a, b] of [a]
         if hasattr(min_left, "__iter__") and hasattr(max_left, "__iter__"):
             # left side is a list
-            lhs = [
-                (a, b, c)
-                for a, b, c in zip(
-                    left_side,
-                    min_left - left_side,
-                    max_left - left_side,
-                )
-            ]
+            left_side = [item for item in left_side]
+            min_left = [item for item in min_left]
+            max_left = [item for item in max_left]
             if hasattr(min_right, "__iter__") and hasattr(max_right, "__iter__"):
                 # right side is a list
-                rhs = [
-                    (a, b, c)
-                    for a, b, c in zip(
-                        right_side,
-                        min_right - right_side,
-                        max_right - right_side,
-                    )
-                ]
-                if len(logger) == 0:
-                    for idx in range(len(lhs)):
+                right_side = [item for item in right_side]
+                min_right = [item for item in min_right]
+                max_right = [item for item in max_right]
+                if len(self._eval_logs) == 0:
+                    for idx in range(len(left_side)):
                         s = (
                             "{"
-                            + str(lhs[idx][0])
+                            + str(left_side[idx])
                             + " - "
-                            + str(rhs[idx][0])
+                            + str(right_side[idx])
                             + " = "
-                            + str(lhs[idx][0] - rhs[idx][0])
+                            + str(left_side[idx] - right_side[idx])
                             + "} "
                             + operator
                             + " "
-                            + "["
-                            + str(lhs[idx][1] - rhs[idx][1])
-                            + ", "
-                            + str(lhs[idx][2] - rhs[idx][2])
-                            + "]"
                         )
-                        logger.append(s)
+                        lower_bound = (
+                            min_left[idx]
+                            - left_side[idx]
+                            - max_right[idx]
+                            + right_side[idx]
+                        )
+                        upper_bound = (
+                            max_left[idx]
+                            - left_side[idx]
+                            - min_right[idx]
+                            + right_side[idx]
+                        )
+                        if lower_bound == upper_bound:
+                            s += "[" + str(lower_bound) + "]"
+                        else:
+                            s += "[" + str(lower_bound) + ", " + str(upper_bound) + "]"
+                        self._eval_logs.append(s)
                 else:
-                    for idx in range(len(lhs)):
+                    for idx in range(len(left_side)):
                         s = (
                             "{"
-                            + str(lhs[idx][0])
+                            + str(left_side[idx])
                             + " - "
-                            + str(rhs[idx][0])
+                            + str(right_side[idx])
                             + " = "
-                            + str(lhs[idx][0] - rhs[idx][0])
+                            + str(left_side[idx] - right_side[idx])
                             + "} "
                             + operator
                             + " "
-                            + "["
-                            + str(lhs[idx][1] - rhs[idx][1])
-                            + ", "
-                            + str(lhs[idx][2] - rhs[idx][2])
-                            + "]"
                         )
-                        logger[idx] += "; " + s
+                        lower_bound = (
+                            min_left[idx]
+                            - left_side[idx]
+                            - max_right[idx]
+                            + right_side[idx]
+                        )
+                        upper_bound = (
+                            max_left[idx]
+                            - left_side[idx]
+                            - min_right[idx]
+                            + right_side[idx]
+                        )
+                        if lower_bound == upper_bound:
+                            s += "[" + str(lower_bound) + "]"
+                        else:
+                            s += "[" + str(lower_bound) + ", " + str(upper_bound) + "]"
+                        self._eval_logs[idx] += "; " + s
             else:
                 # right side is an item
-                if len(logger) == 0:
-                    for idx in range(len(lhs)):
+                if len(self._eval_logs) == 0:
+                    for idx in range(len(left_side)):
                         s = (
                             "{"
-                            + str(lhs[idx][0])
+                            + str(left_side[idx])
                             + " - "
                             + str(right_side)
                             + " = "
-                            + str(lhs[idx][0] - right_side)
+                            + str(left_side[idx] - right_side)
                             + "}"
                         )
-                        logger.append(s)
+                        self._eval_logs.append(s)
                 else:
-                    for idx in range(len(lhs)):
+                    for idx in range(len(left_side)):
                         s = (
                             "{"
-                            + str(lhs[idx][0])
+                            + str(left_side[idx])
                             + " - "
                             + str(right_side)
                             + " = "
-                            + str(lhs[idx][0] - right_side)
+                            + str(left_side[idx] - right_side)
                             + "}"
                         )
-                        logger[idx] += "; " + s
-                for idx in range(len(logger)):
-                    logger[idx] += " " + operator + " "
-                for idx in range(len(lhs)):
+                        self._eval_logs[idx] += "; " + s
+                for idx in range(len(self._eval_logs)):
+                    self._eval_logs[idx] += " " + operator + " "
+                for idx in range(len(left_side)):
                     if operator in ["==", "!="]:
-                        logger[idx] += (
-                            "["
-                            + str(lhs[idx][1] - min_right + right_side)
-                            + ", "
-                            + str(lhs[idx][2] - max_right + right_side)
-                            + "]"
+                        lower_bound = (
+                            min_left[idx] - left_side[idx] - max_right + right_side
                         )
+                        upper_bound = (
+                            max_left[idx] - left_side[idx] - min_right + right_side
+                        )
+                        if lower_bound == upper_bound:
+                            self._eval_logs[idx] += "[" + str(lower_bound) + "]"
+                        else:
+                            self._eval_logs[idx] += (
+                                "[" + str(lower_bound) + ", " + str(upper_bound) + "]"
+                            )
                     elif operator in ["<=", "<"]:
-                        logger[idx] += (
-                            "[" + str(max_right - right_side - lhs[idx][1]) + "]"
-                        )
+                        bound = max_right - right_side - min_left[idx] + left_side[idx]
+                        self._eval_logs[idx] += "[" + str(bound) + "]"
                     elif operator in [">=", ">"]:
-                        logger[idx] += (
-                            "[" + str(min_right - right_side - lhs[idx][2]) + "]"
-                        )
+                        bound = min_right - right_side - max_left[idx] + left_side[idx]
+                        self._eval_logs[idx] += "[" + str(bound) + "]"
         else:
             # left side is an item
             if hasattr(min_right, "__iter__") and hasattr(max_right, "__iter__"):
                 # right side is a list
-                rhs = [
-                    (a, b, c)
-                    for a, b, c in zip(
-                        right_side,
-                        min_right - right_side,
-                        max_right - right_side,
-                    )
-                ]
-                if len(logger) == 0:
-                    for idx in range(len(rhs)):
+                right_side = [item for item in right_side]
+                min_right = [item for item in min_right]
+                max_right = [item for item in max_right]
+                if len(self._eval_logs) == 0:
+                    for idx in range(len(right_side)):
                         s = (
                             "{"
                             + str(left_side)
                             + " - "
-                            + str(rhs[idx][0])
+                            + str(right_side[idx])
                             + " = "
-                            + str(left_side - rhs[idx][0])
+                            + str(left_side - right_side[idx])
                             + "}"
                         )
-                        logger.append(s)
+                        self._eval_logs.append(s)
                 else:
-                    for idx in range(len(rhs)):
+                    for idx in range(len(right_side)):
                         s = (
                             "{"
                             + str(left_side)
                             + " - "
-                            + str(rhs[idx][0])
+                            + str(right_side[idx])
                             + " = "
-                            + str(left_side - rhs[idx][0])
+                            + str(left_side - right_side[idx])
                             + "}"
                         )
-                        logger[idx] += "; " + s
-                for idx in range(len(logger)):
-                    logger[idx] += " " + operator + " "
-                for idx in range(len(rhs)):
+                        self._eval_logs[idx] += "; " + s
+                for idx in range(len(self._eval_logs)):
+                    self._eval_logs[idx] += " " + operator + " "
+                for idx in range(len(right_side)):
                     if operator in ["==", "!="]:
-                        logger[idx] += (
-                            "["
-                            + str(min_left - left_side - rhs[idx][1])
-                            + ", "
-                            + str(max_left - left_side - rhs[idx][2])
-                            + "]"
+                        lower_bound = (
+                            min_left - left_side - max_right[idx] + right_side[idx]
                         )
+                        upper_bound = (
+                            max_left - left_side - min_right[idx] + right_side[idx]
+                        )
+                        if lower_bound == upper_bound:
+                            self._eval_logs[idx] += "[" + str(lower_bound) + "]"
+                        else:
+                            self._eval_logs[idx] += (
+                                "[" + str(lower_bound) + ", " + str(upper_bound) + "]"
+                            )
                     elif operator in ["<=", "<"]:
-                        logger[idx] += (
-                            "[" + str(rhs[idx][2] - min_left + left_side) + "]"
-                        )
+                        bound = max_right[idx] - right_side[idx] - min_left + left_side
+                        self._eval_logs[idx] += "[" + str(bound) + "]"
                     elif operator in [">=", ">"]:
-                        logger[idx] += (
-                            "[" + str(rhs[idx][1] - max_left + left_side) + "]"
-                        )
+                        bound = min_right[idx] - right_side[idx] - max_left + left_side
+                        self._eval_logs[idx] += "[" + str(bound) + "]"
 
             else:
                 # right side is a item
-                logger = (
+                self._eval_logs += (
                     "{"
                     + str(left_side)
                     + " - "
@@ -1091,23 +864,19 @@ class CodeEvaluator:
                     + str(left_side - right_side)
                     + "}"
                 )
-                logger += " " + operator + " "
+                self._eval_logs += " " + operator + " "
                 if operator in ["==", "!="]:
-                    logger += (
-                        "["
-                        + str(min_left - min_left - min_right + right_side)
-                        + ", "
-                        + str(max_left - max_left - max_right + right_side)
-                        + "]"
+                    lower_bound = min_left - left_side - max_right + right_side
+                    upper_bound = max_left - left_side - min_right + right_side
+                    self._eval_logs += (
+                        "[" + str(lower_bound) + ", " + str(upper_bound) + "]"
                     )
                 elif operator in ["<=", "<"]:
-                    logger += (
-                        "[" + str(max_right - right_side - min_left + left_side) + "]"
-                    )
+                    bound = max_right - right_side - min_left + left_side
+                    self._eval_logs += "[" + str(bound) + "]"
                 elif operator in [">=", ">"]:
-                    logger += (
-                        "[" + str(min_right - right_side - max_left + left_side) + "]"
-                    )
+                    bound = min_right - right_side - max_left + left_side
+                    self._eval_logs += "[" + str(bound) + "]"
 
     def set_params(self, params):
         """
@@ -1218,12 +987,7 @@ class CodeEvaluator:
                 self._mean_logs = []
                 self._std_logs = []
                 self._quantile_logs = []
-                self._eq_logs = []
-                self._ne_logs = []
-                self._ge_logs = []
-                self._le_logs = []
-                self._gt_logs = []
-                self._lt_logs = []
+                self._eval_logs = []
                 if key == "X":
                     logs += "if ("
                     logs_added = False
@@ -1242,19 +1006,11 @@ class CodeEvaluator:
                     if len(self._quantile_logs) > 0:
                         log.append("; ".join(self._quantile_logs))
                     # put logs in pd.Series as a strings
-                    for item in [
-                        self._eq_logs,
-                        self._ne_logs,
-                        self._ge_logs,
-                        self._le_logs,
-                        self._gt_logs,
-                        self._lt_logs,
-                    ]:
-                        if len(item) > 0:
-                            if logs_added:
-                                logs += "; "
-                            logs += item
-                            logs_added = True
+                    if len(self._eval_logs) > 0:
+                        if logs_added:
+                            logs += "; "
+                        logs += self._eval_logs
+                        logs_added = True
                     if len(log) > 0:
                         if logs_added:
                             logs += "; "